:recycle: #459 - refactor: combine hooks directories
