+++ frontend/src/hooks/index.ts
@@ -1,6 +1,16 @@
 export * from "./useAuditLog";
+export * from "./useCombinedSearchParams";
+export * from "./useFields";
+export * from "./useFilter";
 export * from "./useLatestReviewResponse";
+export * from "./usePage";
 export * from "./usePoll";
 export * from "./useReviewers";
+export * from "./useSelectielijstKlasseChoices";
+export * from "./useSort";
 export * from "./useSubmitAction";
 export * from "./useWhoAmI";
+export * from "./useZaakReviewStatusBadges";
+export * from "./useZaakReviewStatuses";
+export * from "./useZaakSelection";
+export * from "./useZaaktypeChoices";
