+++ frontend/src/hooks/useCombinedSearchParams.ts
