+++ frontend/src/hooks/useFields.test.ts
@@ -1,8 +1,8 @@
 import { TypedField } from "@maykin-ui/admin-ui";
 import { act, renderHook, waitFor } from "@testing-library/react";
 
-import { selectieLijstKlasseFactory as mockSelectieLijstKlasseFactory } from "../../../fixtures/selectieLijstKlasseChoices";
-import * as fieldSelection from "../../../lib/fieldSelection/fieldSelection";
+import { selectieLijstKlasseFactory as mockSelectieLijstKlasseFactory } from "../fixtures/selectieLijstKlasseChoices";
+import * as fieldSelection from "../lib/fieldSelection/fieldSelection";
 import { useFields } from "./useFields";
 
 jest.mock("./useSelectielijstKlasseChoices", () => ({
