--- frontend/src/pages/destructionlist/hooks/index.ts
@@ -1,10 +1,0 @@
-export * from "./useCombinedSearchParams";
-export * from "./useFields";
-export * from "./useFilter";
-export * from "./usePage";
-export * from "./useSelectielijstKlasseChoices";
-export * from "./useSort";
-export * from "./useZaakReviewStatusBadges";
-export * from "./useZaakReviewStatuses";
-export * from "./useZaakSelection";
-export * from "./useZaaktypeChoices";
